requirements.txt
numpy
matplotlib
scipy
pyyaml
visa
PyQt5
configparser
pandas


python >= 3.0